Bug fix:  When new users sign up using the red "Sign in with email" button, the displayName was not available as soon as the auth event was detected.  The displayName doesn't become available until just a little bit later in CenterViewController.authUI()
CenterViewController.authUI() - This is where user.displayName is available.  So this is where we set the name attribute of the TPUser object and then we update the database on the very next line.

LimboViewController.changed(name: String) - added an implementation to this method that has, up to now, just been a loop.  Here's what we're doing here...  The LimboViewController checks for missing name and missing email on the user.  If either are missing, this view controller displays the MissingInformationView, a view that displays a name field and an email field so that the user could manually enter his name or email (whatever was missing).  But re: the name attribute, it CAN'T be missing if the user signed up using the "Sign in with email" button because that button presents you with a screen that requires a name to be supplied.  So up to now, users were having to enter their names twice - once as part of normal screen flow when creating an account, but also a second time in this MissingInformationView.  Now that we figured out where user.displayName becomes available - CenterViewController.authUI() - we can set the name there, update the database and fire the AccountStatusEventListener fireChanged(name: String) which ends up calling LimboViewController.changed (because LimboViewController is an AccountStatusEventListener).  So LimboViewController.changed() removes the missingInformationView if it's being displayed, and in its place, displays the LimboView

TPUser.set(name: String) - simply added a call to fireChange(name: String) so that all AccountStatusEventListeners get notified.  As the comments in this function indicate, this fireChange() call really exists to make sure that we don't display the MissingInformationView